requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
distlib==0.3.9
filelock==3.16.1
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.1.2
opencv-python==4.10.0.84
platformdirs==4.3.6
pyserial==3.5
scikit-learn==1.5.2
scipy==1.14.1
threadpoolctl==3.5.0
virtualenv==20.26.6
Werkzeug==3.1.3
